Improvement on Code
Improvement the reusability of the code and deleting some duplicate code
